Refactor tests: extract common parts into abstract base class
